+++ R/parallel_funs.R
@@ -26,8 +26,8 @@ kmbayes_parallel <- function(nchains=4, ...) {
   #' X <- dat$X
   #' set.seed(111)
   #' Sys.setenv(R_FUTURE_SUPPORTSMULTICORE_UNSTABLE="quiet")
-  #' future::plan(strategy = future::multiprocess)
-  #' fitkm.list <- kmbayes_parallel(nchains=4, y = y, Z = Z, X = X, iter = 5000,
+  #' future::plan(strategy = future::multiprocess, workers=2)
+  #' fitkm.list <- kmbayes_parallel(nchains=2, y = y, Z = Z, X = X, iter = 5000,
   #'   verbose = FALSE, varsel = TRUE)
   #' closeAllConnections()
   #' }
@@ -79,9 +79,9 @@ kmbayes_combine <- function(fitkm.list, burnin=0, reorder=TRUE) {
   #' X <- dat$X
   #' set.seed(111)
   #' Sys.setenv(R_FUTURE_SUPPORTSMULTICORE_UNSTABLE="quiet")
-  #' future::plan(strategy = future::multiprocess)
+  #' future::plan(strategy = future::multiprocess, workers=2)
   #' # run 4 parallel Markov chains
-  #' fitkm.list <- kmbayes_parallel(nchains=4, y = y, Z = Z, X = X, iter = 5000,
+  #' fitkm.list <- kmbayes_parallel(nchains=2, y = y, Z = Z, X = X, iter = 5000,
   #'   verbose = FALSE, varsel = TRUE)
   #' bigkm = kmbayes_combine(fitkm.list)
   #' ests = ExtractEsts(bigkm)