+++ tests/test_parallel.R
@@ -7,7 +7,7 @@ Z <- dat$Z
 X <- dat$X
 set.seed(111)
 Sys.setenv(R_FUTURE_SUPPORTSMULTICORE_UNSTABLE="quiet")
-future::plan(strategy = future::multiprocess)
+future::plan(strategy = future::sequential, workers=2)
 fitkm.list <- kmbayes_parallel(nchains=2, y = y, Z = Z, X = X, iter = 10,
                                verbose = FALSE, varsel = TRUE)
 PredictorResponseBivar_parallel(fitkm.list)
@@ -19,3 +19,7 @@ SamplePred_parallel(fitkm.list)
 comb_bkmrfits(fitkm.list)
 kmbayes_diag(fitkm.list)
 closeAllConnections()
+
+future::plan(strategy = future::multiprocess, workers=2)
+fitkm.list <- kmbayes_parallel(nchains=2, y = y, Z = Z, X = X, iter = 10,
+                               verbose = FALSE, varsel = TRUE)