+++ NEWS.md
@@ -1,3 +1,12 @@
+# `bkmrhat` v0.3.7
+
+## Major changes
+- Added parallel diagnostics and re-packaged bkmr package functions for parallel chains
+- Added "continue" functions to continue fitting from an existing bkmr fit
+
+## Bug fixes
+- N/A
+
 # `bkmrhat` v0.1.16
 
 ## Major changes